I think this fixes the calendar update problem...
The spec requires us to update a LAST-MODIFIED timestamp when we change
events. Google is pretty aggressive about this and won't update events
if we don't do this, even if it pulls down the updated ICS file.

Then separately, we were using DTSTAMP wrong -- it's the event's
creation time, and Google looks at both it and LAST-MODIFIED to decide
if an event has been updated. Since we screwed this up by using the
event's start time as DTSTAMP, Google's cache currently thinks the
LAST-MODIFIED for all the existing events is in the future. To work
around Google's cache this one time, increment the SEQUENCE number to
indicate the events have been edited.